Creacion de documentos para separar la funcion de pipes, junto a un arreglo de argumentos
