+++ Comando_pipes.h
@@ -1,5 +1,5 @@
-#ifndef COMANDOMIPROF_H
-#define COMANDOMIPROF_H
+#ifndef COMANDOPIPE_H
+#define COMANDOPIPE_H
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -7,5 +7,5 @@
 #include <sys/wait.h>
 #include <fcntl.h>
 
-void comando_miprof(char **comandos);
+void ejecutar_pipes(char *comandos[], int num_comandos);
 #endif
