+++ Shell.c
@@ -5,6 +5,7 @@
 #include <sys/wait.h>
 #include <fcntl.h>
 #include "Comando_personalizado.h"
+#include "Comando_pipes.h"
 
 void ejecutar_pipes(char *comandos[], int num_comandos);
 
@@ -52,13 +53,11 @@ int main(int argc, char *argv[])
             break;
         }
 
-    
         if (strcmp(comandos[0], "miprof") == 0)
         {
             manejar_miprof(comandos);
             continue;
         }
-    
 
         // Para saber si ejecutar comando simple o pipes
         if (num_comandos == 1)
@@ -102,80 +101,4 @@ int main(int argc, char *argv[])
     }
 
     return 0;
-}
-
-void ejecutar_pipes(char *comandos[], int num_comandos)
-{
-    int i;
-    int in_fd = 0; // stdin
-    int fd[2];
-    pid_t pid;
-
-    for (i = 0; i < num_comandos; i++)
-    {
-        // crear pipes para todo menos para el último comando
-        if (i < num_comandos - 1)
-        {
-            if (pipe(fd) < 0)
-            {
-                perror("error al ejecutar pipe");
-                return;
-            }
-        }
-
-        pid = fork();
-        if (pid == 0)
-        { // hijo
-            if (i > 0)
-            {
-                // Redirige entrada desde el pipe anterior
-                dup2(in_fd, STDIN_FILENO);
-                close(in_fd);
-            }
-            if (i < num_comandos - 1)
-            {
-                // Redirige entrada al pipe siguiente
-                dup2(fd[1], STDOUT_FILENO);
-                close(fd[1]);
-                close(fd[0]);
-            }
-            // ===PARSEAR===
-            char *args[256];
-            char *token = strtok(comandos[i], " ");
-            int j = 0;
-            while (token != NULL)
-            {
-                args[j++] = token;
-                token = strtok(NULL, " ");
-            }
-            args[j] = NULL;
-
-            execvp(args[0], args);
-            perror("error execvp");
-            exit(1);
-        }
-        else if (pid > 0)
-        { // PADRE
-            if (i > 0)
-            {
-                close(in_fd);
-            }
-            if (i < num_comandos - 1)
-            {
-                close(fd[1]);
-                in_fd = fd[0]; // entrada para el proximo cmdo
-            }
-        }
-        else
-        {
-            perror("Error en el fork");
-            return;
-        }
-    }
-
-    // para esperar todos los procesos hijos:
-    for (i = 0; i < num_comandos; i++)
-    {
-        wait(NULL);
-    }
 }
